line2

+++ app/src/main/java/uz/koinot/stadion/data/model/Dashboard.kt
@@ -2,5 +2,5 @@ package uz.koinot.stadion.data.model
 
 data class Dashboard(
     var day:String,
-    var benefit:Float,
+    var benefit:Double,
 )
